requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
statsmodels
seaborn
matplotlib
prophet
xgboost
chardet
shap
google-generativeai
